requirements.txt
alembic==1.4.2
astroid==2.3.3
Babel==2.8.0
blinker==1.4
cffi==1.14.0
click==7.1.1
cryptography==2.9
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.0.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.20
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.16
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.2.1
